Fix to ARM disasm

+++ com.oracle.max.asmdis/src/com/sun/max/asm/dis/arm/ARMDisassembler.java
@@ -38,8 +38,8 @@ import com.sun.max.lang.*;
  */
 public class ARMDisassembler extends RiscDisassembler{
 
-    public ARMDisassembler(long startAddress, InlineDataDecoder inlineDataDecoder) {
-        super(new Immediate32Argument((int) startAddress), ARMAssembly.ASSEMBLY, Endianness.BIG, inlineDataDecoder);
+    public ARMDisassembler(int startAddress, InlineDataDecoder inlineDataDecoder) {
+        super(new Immediate32Argument(startAddress), ARMAssembly.ASSEMBLY, Endianness.BIG, inlineDataDecoder);
     }
 
     @Override
